* Deleting your own alias or implication no longer sends a dmail * Implements #344: Withdraw alias/implication suggestions * Tested OK
git-svn-id: svn://donmai.us/danbooru/trunk@1615 9af9e794-c952-11db-92b5-8e906a41d509

+++ app/controllers/tag_alias_controller.rb
@@ -1,6 +1,5 @@
 class TagAliasController < ApplicationController
   layout "default"
-  before_filter :admin_only, :only => [:update]
   before_filter :member_only, :only => [:create]
   verify :method => :post, :only => [:create, :update]
 
@@ -39,16 +38,24 @@ class TagAliasController < ApplicationController
 
     case params[:commit]
     when "Delete"
-      ids.each {|x| TagAlias.find(x).destroy_and_notify(@current_user, params[:reason])}
+      if @current_user.is_admin? || ids.all? {|x| ta = TagAlias.find(x) ; ta.is_pending? && ta.creator_id == @current_user.id}
+        ids.each {|x| TagAlias.find(x).destroy_and_notify(@current_user, params[:reason])}
       
-      flash[:notice] = "Tag aliases deleted"
-      redirect_to :action => "index"
+        flash[:notice] = "Tag aliases deleted"
+        redirect_to :action => "index"
+      else
+        access_denied
+      end
 
     when "Approve"
-      ids.each {|x| TagAlias.find(x).approve(@current_user.id, request.remote_ip)}
+      if @current_user.is_admin?
+        ids.each {|x| TagAlias.find(x).approve(@current_user.id, request.remote_ip)}
 
-      flash[:notice] = "Tag aliases approved"
-      redirect_to :action => "index"
+        flash[:notice] = "Tag aliases approved"
+        redirect_to :action => "index"
+      else
+        access_denied
+      end
     end
   end
 end
